requirements.txt
telethon==1.42.0
python-telegram-bot==20.7
fastapi
uvicorn
